requirements.txt
amqp==5.1.1
asgiref==3.5.2
attrs==22.2.0
autobahn==22.12.1
Automat==22.10.0
billiard==3.6.4.0
boto3==1.26.50
botocore==1.29.50
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
constantly==15.1.0
cryptography==39.0.0
daphne==4.0.0
Django==4.1.4
django-cors-headers==3.13.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jmespath==1.0.1
kombu==5.2.4
prompt-toolkit==3.0.36
psycopg2==2.9.5
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.0.0
pyotp==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.6
requests==2.28.2
s3transfer==0.6.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
twilio==7.16.1
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
zope.interface==5.5.2
